+++ src/Sulu/Bundle/Sales/OrderBundle/Resources/public/js/components/orders/components/form/main.js
@@ -13,9 +13,10 @@ define([
     'sulusalesorder/util/header',
     'sulusalescore/util/helper',
     'sulucontact/models/account',
+    'sulucontact/models/contact',
     'config',
     'widget-groups'
-], function(Sidebar, OrderStatus, HeaderUtil, CoreHelper, Account, Config, WidgetGroups) {
+], function(Sidebar, OrderStatus, HeaderUtil, CoreHelper, Account, Contact, Config, WidgetGroups) {
 
     'use strict';
 
@@ -37,26 +38,42 @@ define([
          */
         EVENT_SET_OPTIONS_DATA = namespace + 'set-options-data',
 
+        CUSTOMER_TYPE = {
+            ORGANIZATION: 1,
+            PRIVATE_PERSON: 2
+        },
+
         constants = {
             accountContactsUrl: '/admin/api/accounts/<%= id %>/contacts?flat=true',
             accountAddressesUrl: '/admin/api/accounts/<%= id %>/addresses',
-            accountInputId: '#account-input',
+            customerAccountId: 'js-customer-account',
+            customerContactId: 'js-customer-contact',
             deliveryAddressInstanceName: 'delivery-address',
             billingAddressInstanceName: 'billing-address',
             currencySelectInstanceName: 'currency-select',
-            currencySelectSelector: '#currency-select',
+            customerTypeSelectInstanceName: 'customer-type-select',
             itemTableInstanceName: 'order-items',
-            itemTableSelector: '#order-items',
             paymentTermsInstanceName: 'payment-terms',
             deliveryTermsInstanceName: 'delivery-terms',
-            contactSelectId: '#contact-select',
+            contactSelectInstanceName: 'contact-select',
             validateWarningTranslation: 'form.validation-warning',
             translationShippingFailed: 'salescore.shipping-failed',
             autocompleteLimit: 20
         },
 
+        selectors = {
+            customerAutocompleteContainer: '#js-customer-autocomplete-container',
+            customerAccount: '#' + constants.customerAccountId,
+            customerContact: '#' + constants.customerContactId,
+            currencySelect: '#js-currency-select',
+            deliveryAddress: '#js-' + constants.deliveryAddressInstanceName,
+            billingAddress: '#js-' + constants.billingAddressInstanceName,
+            itemTable: '#js-order-items',
+            taxFree: '#js-tax-free'
+        },
+
         /**
-         * get status of current order
+         * Get status of current order.
          */
         getOrderStatusId = function() {
             return (!!this.options.data && !!this.options.data.status) ?
@@ -66,8 +83,8 @@ define([
         /**
          * Sets specific data to options.
          *
-         * @param {String} key Where to set data (this.options[key])
-         * @param {Mixed} optionData Data to set onto options
+         * @param {String} key Where to set data (this.options[key]).
+         * @param {Mixed} optionData Data to set onto options.
          */
         setOptionsData = function(key, optionData) {
             this.options[key] = optionData;
@@ -76,28 +93,55 @@ define([
         bindCustomEvents = function() {
             this.sandbox.on(EVENT_SET_OPTIONS_DATA, setOptionsData.bind(this));
 
-            this.sandbox.on('husky.auto-complete.' + this.accountInstanceName + '.initialized', function() {
+            // Auto complete customer account.
+            this.sandbox.on('husky.auto-complete.' + this.customerAccountInstanceName + '.initialized', function() {
                 if (!this.isEditable) {
-                    this.sandbox.dom.attr(this.$find(constants.accountInputId + ' input'), 'disabled', 'disabled');
+                    this.sandbox.dom.attr(this.$find(selectors.customerAccount + ' input'), 'disabled', 'disabled');
                 }
                 this.dfdAutoCompleteInitialized.resolve();
             }, this);
 
-            this.sandbox.on('husky.auto-complete.' + this.accountInstanceName + '.selection-removed', accountChangedListener.bind(this));
+            this.sandbox.on(
+                'husky.auto-complete.' + this.customerAccountInstanceName + '.select',
+                accountChangedListener.bind(this)
+            );
 
-            // contact saved
+            this.sandbox.on(
+                'husky.auto-complete.' + this.customerAccountInstanceName + '.selection-removed',
+                accountChangedListener.bind(this)
+            );
+
+            // Auto complete customer contact.
+            this.sandbox.on('husky.auto-complete.' + this.customerContactInstanceName + '.initialized', function() {
+                if (!this.isEditable) {
+                    this.sandbox.dom.attr(this.$find(selectors.customerContact + ' input'), 'disabled', 'disabled');
+                }
+                this.dfdAutoCompleteInitialized.resolve();
+            }, this);
+
+            this.sandbox.on(
+                'husky.auto-complete.' + this.customerContactInstanceName + '.select',
+                contactChangedListener.bind(this)
+            );
+
+            this.sandbox.on(
+                'husky.auto-complete.' + this.customerContactInstanceName + '.selection-removed',
+                contactChangedListener.bind(this)
+            );
+
+            // Contact saved.
             this.sandbox.on('sulu.salesorder.order.saved', function(data) {
                 this.options.data = data;
                 setSaved.call(this, true);
                 this.dfdFormSaved.resolve();
             }, this);
 
-            // contact save
+            // Contact save.
             this.sandbox.on('sulu.toolbar.save', function() {
                 this.submit();
             }, this);
 
-            // back to list
+            // Back to list.
             this.sandbox.on('sulu.header.back', function() {
                 this.sandbox.emit('sulu.salesorder.orders.list');
             }, this);
@@ -110,27 +154,37 @@ define([
                 this.dfdOrderDate.resolve();
             }, this);
 
-            this.sandbox.on('husky.auto-complete.' + this.accountInstanceName + '.select', accountChangedListener.bind(this));
+            this.sandbox.on(
+                'sulu.editable-data-row.' + constants.deliveryAddressInstanceName + '.initialized',
+                function() {
+                    this.dfdDeliveryAddressInitialized.resolve();
+                }.bind(this)
+            );
 
-            this.sandbox.on('sulu.editable-data-row.' + constants.deliveryAddressInstanceName + '.initialized', function() {
-                this.dfdDeliveryAddressInitialized.resolve();
-            }.bind(this));
+            this.sandbox.on(
+                'sulu.editable-data-row.' + constants.billingAddressInstanceName + '.initialized',
+                function() {
+                    this.dfdInvoiceAddressInitialized.resolve();
+                }.bind(this)
+            );
 
-            this.sandbox.on('sulu.editable-data-row.' + constants.billingAddressInstanceName + '.initialized', function() {
-                this.dfdInvoiceAddressInitialized.resolve();
-            }.bind(this));
+            this.sandbox.on(
+                'sulu.editable-data-row.address-view.' + constants.deliveryAddressInstanceName + '.changed',
+                function(data) {
+                    this.options.data.deliveryAddress = data;
+                    setFormData.call(this, this.options.data);
+                    changeHandler.call(this);
+                }.bind(this)
+            );
 
-            this.sandbox.on('sulu.editable-data-row.address-view.' + constants.deliveryAddressInstanceName + '.changed', function(data) {
-                this.options.data.deliveryAddress = data;
-                setFormData.call(this, this.options.data);
-                changeHandler.call(this);
-            }.bind(this));
-
-            this.sandbox.on('sulu.editable-data-row.address-view.' + constants.billingAddressInstanceName + '.changed', function(data) {
-                this.options.data.invoiceAddress = data;
-                setFormData.call(this, this.options.data);
-                changeHandler.call(this);
-            }.bind(this));
+            this.sandbox.on(
+                'sulu.editable-data-row.address-view.' + constants.billingAddressInstanceName + '.changed',
+                function(data) {
+                    this.options.data.invoiceAddress = data;
+                    setFormData.call(this, this.options.data);
+                    changeHandler.call(this);
+                }.bind(this)
+            );
 
             this.sandbox.on('husky.select.' + constants.currencySelectInstanceName + '.selected.item', function(data) {
                 this.sandbox.emit('sulu.item-table.' + constants.itemTableInstanceName + '.change-currency', data);
@@ -140,14 +194,37 @@ define([
             this.sandbox.on('sulu.salesorder.set-customer-id', function(customerId) {
                 this.customerId = customerId;
             }, this);
+
+            this.sandbox.on('husky.select.' + constants.contactSelectInstanceName + '.initialize', function() {
+                this.dfdContactInitialized.resolve();
+            }, this);
+
+            this.sandbox.on('husky.select.' + constants.customerTypeSelectInstanceName + '.initialize', function() {
+                this.dfdCustomerTypeInitialized.resolve();
+            }, this);
+
+            this.sandbox.on(
+                'husky.select.' + constants.customerTypeSelectInstanceName + '.selected.item',
+                function(customerTypeId) {
+                    customerTypeChangedHandler.call(this, parseInt(customerTypeId));
+                },
+                this
+            );
         },
 
         bindDomEvents = function() {
-            this.sandbox.dom.on(this.$el, 'click', onTaxfreeClicked.bind(this), '#tax-free');
+            this.sandbox.dom.on(
+                this.$el,
+                'click',
+                onTaxfreeClicked.bind(this),
+                selectors.taxFree
+            );
         },
 
         /**
-         * @param {event} gets called when checkbox is triggered
+         * Event gets called when checkbox is triggered.
+         *
+         * @param {Object} event
          */
         onTaxfreeClicked = function(event) {
             var taxfree = $(event.currentTarget).is(':checked');
@@ -155,8 +232,9 @@ define([
         },
 
         /**
-         * set saved
-         * @param {Bool} saved Defines if saved state should be shown
+         * Set saved method.
+         *
+         * @param {Boolean} saved Defines if saved state should be shown.
          */
         setSaved = function(saved) {
             if (saved !== this.saved) {
@@ -169,119 +247,256 @@ define([
             this.saved = saved;
         },
 
+        /**
+         * @param {Object} data
+         */
         initForm = function(data) {
             var formObject = this.sandbox.form.create(form);
             formObject.initialized.then(function() {
                 setFormData.call(this, data, true);
-                // TODO: resolve that form is set
                 startFormComponents.call(this, data);
             }.bind(this));
         },
 
+        /**
+         * @param {Object} data
+         */
         setFormData = function(data) {
-            // add collection filters to form
+            // Add collection filters to form.
             this.sandbox.form.setData(form, data).then(function() {
                 this.accountId = getAccountId.call(this);
+                this.contactId = getContactId.call(this);
             }.bind(this)).fail(function(error) {
                 this.sandbox.logger.error("An error occured when setting data!", error);
             }.bind(this));
         },
 
         /**
-         * start components form components
-         * @param data
+         * Start components form components.
+         *
+         * @param {Object} data
          */
         startFormComponents = function(data) {
             this.sandbox.start(form);
 
-            if (!!data.customerAccount && !!data.customerAccount.id) {
-                initSelectsByAccountId.call(this, data.customerAccount.id, data);
+            if (!!data && !data.customerAccount && !!data.customerContact && !!data.customerContact.id) {
+                startCustomerContactAutocomplete.call(this, data);
+
+                return;
             }
 
+            startCustomerAccountAutocomplete.call(this, data);
+        },
+
+        /**
+         * @param {Object} data
+         */
+        startCustomerAccountAutocomplete = function(data) {
+            // Stop other autocomplete element.
+            this.sandbox.stop(selectors.customerContact);
+
+            // Create dom element.
+            var $element = this.sandbox.dom.createElement('<div id="' + constants.customerAccountId + '" />');
+            this.sandbox.dom.append(this.$find(selectors.customerAutocompleteContainer), $element);
+
             var options = Config.get('sulucontact.components.autocomplete.default.account');
-            options.el = constants.accountInputId;
-            options.value = !!data.customerAccount ? data.customerAccount : '';
-            options.instanceName = this.accountInstanceName;
+            options.el = selectors.customerAccount;
+            options.value = !!data && !!data.customerAccount ? data.customerAccount : '';
+            options.instanceName = this.customerAccountInstanceName;
             options.remoteUrl += '&type=' + this.customerId + '&limit=' + constants.autocompleteLimit;
             options.limit = constants.autocompleteLimit;
 
-            // starts form components
+            // Starts form components.
             this.sandbox.start([
                 {
                     name: 'auto-complete@husky',
                     options: options
                 }
             ]);
+
+            if (!!data && !!data.customerAccount && !!data.customerAccount.id) {
+                initSelectsByAccountId.call(this, data.customerAccount.id, data);
+            }
         },
 
         /**
-         * returns id of currently set account
-         * @returns string
+         * @param {Object} data
+         */
+        startCustomerContactAutocomplete = function(data) {
+            // Stop other autocomplete element.
+            this.sandbox.stop(selectors.customerAccount);
+
+            // Create dom element.
+            var $element = this.sandbox.dom.createElement('<div id="' + constants.customerContactId + '" />');
+            this.sandbox.dom.append(this.$find(selectors.customerAutocompleteContainer), $element);
+
+            var customerContactOptions = Config.get('sulucontact.components.autocomplete.default.contact');
+            customerContactOptions.el = selectors.customerContact;
+            customerContactOptions.value = !!data && !!data.customerContact ? data.customerContact : '';
+            customerContactOptions.instanceName = this.customerContactInstanceName;
+            customerContactOptions.limit = constants.autocompleteLimit;
+
+            // Starts form components.
+            this.sandbox.start([
+                {
+                    name: 'auto-complete@husky',
+                    options: customerContactOptions
+                }
+            ]);
+
+            if (!!data && !!data.customerContact && !!data.customerContact.id) {
+                initSelectsByContact.call(this, data.customerContact, data);
+            }
+        },
+
+        /**
+         * Returns id of currently set account.
+         *
+         * @returns {String}
          */
         getAccountId = function() {
-            return this.sandbox.dom.attr(constants.accountInputId, 'data-id');
+            return this.sandbox.dom.attr(selectors.customerAccount, 'data-id');
         },
 
         /**
-         * init contact select
-         * @param data
-         * @param preselectedElements
+         * @returns {String}
+         */
+        getContactId = function() {
+            return this.sandbox.dom.attr(selectors.customerContact, 'data-id');
+        },
+
+        /**
+         * Init customer type select.
+         *
+         * @param {Number} customerTypeId
+         */
+        initCustomerTypeSelect = function(customerTypeId) {
+            this.customerTypeId = customerTypeId;
+
+            this.sandbox.data.when(this.dfdCustomerTypeInitialized).then(function() {
+                this.sandbox.emit(
+                    'husky.select.customer-type-select.update',
+                    this.options.customerTypes,
+                    [customerTypeId]
+                );
+            }.bind(this));
+        },
+
+        /**
+         * Init contact select.
+         *
+         * @param {Array} data
+         * @param {Array} preselectedElements
          */
         initContactSelect = function(data, preselectedElements) {
-
             preselectedElements = preselectedElements || [];
 
-            this.sandbox.emit('husky.select.contact-select.update', data, preselectedElements);
+            this.sandbox.data.when(this.dfdContactInitialized).then(function() {
+                this.sandbox.emit('husky.select.contact-select.update', data, preselectedElements);
+            }.bind(this));
         },
 
         /**
-         * init address select
-         * @param data
-         * @param instanceName
-         * @param preselectedElement
+         * Init address select.
+         *
+         * @param {Array} data
+         * @param {String} instanceName
+         * @param {Array} preselectedElement
          */
-        initAddressComponents = function(data, instanceName, preselectedElement) {
+        initAddressComponent = function(data, instanceName, preselectedElement) {
             this.sandbox.emit('sulu.editable-data-row.' + instanceName + '.data.update', data, preselectedElement);
-            this.$find('#' + constants.deliveryAddressInstanceName).removeClass('disabled-button');
-            this.$find('#' + constants.billingAddressInstanceName).removeClass('disabled-button');
+            this.$find(selectors.deliveryAddress).removeClass('disabled-button');
+            this.$find(selectors.billingAddress).removeClass('disabled-button');
         },
 
         /**
-         * set value of editable data-row
-         * @param instanceName
-         * @param value
+         * Init delivery and invoice address component.
+         *
+         * @param {Object} orderData
+         * @param {Array} addressData
+         */
+        initAddressComponents = function(orderData, addressData)
+        {
+            // When an address is already selected, the selected address should be used
+            // otherwise the first delivery / payment address found will be used.
+            var preselect = null;
+            if (!orderData || !orderData.deliveryAddress) {
+                preselect = findAddressWherePropertyIs.call(this, addressData, 'deliveryAddress', true);
+
+                // When no delivery address is found the first address will be used.
+                if (!preselect && addressData.length > 0) {
+                    preselect = addressData[0];
+                }
+            } else {
+                preselect = orderData.deliveryAddress;
+            }
+            this.sandbox.data.when(this.dfdDeliveryAddressInitialized).then(function() {
+                initAddressComponent.call(this, addressData, constants.deliveryAddressInstanceName, preselect);
+                setSettingsOverlayAdresses.call(this, addressData, preselect);
+                this.options.data.deliveryAddress = preselect;
+            }.bind(this));
+
+            preselect = null;
+            if (!orderData || !orderData.invoiceAddress) {
+                preselect = findAddressWherePropertyIs.call(this, addressData, 'billingAddress', true);
+
+                // When no invoice address is found the first address will be used.
+                if (!preselect && addressData.length > 0) {
+                    preselect = addressData[0];
+                }
+            } else {
+                preselect = orderData.invoiceAddress;
+            }
+
+            this.sandbox.data.when(this.dfdInvoiceAddressInitialized).then(function() {
+                initAddressComponent.call(this, addressData, constants.billingAddressInstanceName, preselect);
+                this.options.data.invoiceAddress = preselect;
+            }.bind(this));
+        },
+
+        /**
+         * Set value of editable data-row.
+         *
+         * @param {String} instanceName
+         * @param {Array} value
          */
         setValueOfEditableDataRow = function(instanceName, value) {
             this.sandbox.emit('sulu.editable-data-row.' + instanceName + '.set-value', value);
         },
 
         /**
-         * set addresses for settings overlay
-         * @param addresses
-         * @param preselect
+         * Set addresses for settings overlay.
+         *
+         * @param {Array} addresses
+         * @param {Array} preselect
          */
         setSettingsOverlayAdresses = function(addresses, preselect) {
-            this.sandbox.emit('sulu.item-table.' + constants.itemTableInstanceName + '.set-addresses', addresses, preselect);
+            this.sandbox.emit(
+                'sulu.item-table.' + constants.itemTableInstanceName + '.set-addresses',
+                addresses,
+                preselect
+            );
         },
 
         /**
-         * called when account auto-complete is changed
-         * @param event
+         * Called when account auto-complete is changed.
+         *
+         * @param {Object} event
          */
         accountChangedListener = function(event) {
-            var id = event.id || null;
+            var id = (event && event.id) ? event.id : null;
 
-            // if account has been changed
+            // If account has been changed.
             if (id !== this.accountId) {
                 this.accountId = id;
 
                 if (id) {
-                    // load contacts of account
+                    // Load contacts of account.
                     initSelectsByAccountId.call(this, id);
                 } else {
                     initContactSelect.call(this, []);
-                    initAddressComponents.call(this, [], constants.deliveryAddressInstanceName);
-                    initAddressComponents.call(this, [], constants.billingAddressInstanceName);
+                    initAddressComponent.call(this, [], constants.deliveryAddressInstanceName);
+                    initAddressComponent.call(this, [], constants.billingAddressInstanceName);
 
                     setSettingsOverlayAdresses.call(this, []);
                 }
@@ -289,21 +504,100 @@ define([
         },
 
         /**
-         * called when headerbar should be saveable
+         * Called when contact auto-complete is changed.
+         *
+         * @param {Object} contact
+         */
+        contactChangedListener = function(contact) {
+            var id = (contact && contact.id) ? contact.id : null;
+
+            // If contact has been changed.
+            if (id !== this.contactId) {
+                this.contactId = id;
+
+                if (id) {
+                    initSelectsByContact.call(this, contact);
+                } else {
+                    initContactSelect.call(this, []);
+                    initAddressComponent.call(this, [], constants.deliveryAddressInstanceName);
+                    initAddressComponent.call(this, [], constants.billingAddressInstanceName);
+
+                    setSettingsOverlayAdresses.call(this, []);
+                }
+            }
+        },
+
+        /**
+         * @param {Number} customerTypeId
+         */
+        customerTypeChangedHandler = function(customerTypeId) {
+            // If type has changed, load the correct autocomplete element.
+            if (this.customerTypeId != customerTypeId) {
+                this.customerTypeId = customerTypeId;
+                switch (this.customerTypeId) {
+                    case CUSTOMER_TYPE.ORGANIZATION:
+                        startCustomerAccountAutocomplete.call(this);
+                        accountChangedListener.call(this, null);
+                        break;
+                    case CUSTOMER_TYPE.PRIVATE_PERSON:
+                        startCustomerContactAutocomplete.call(this);
+                        contactChangedListener.call(this, null);
+                        break;
+                }
+            }
+        },
+
+        /**
+         * Called when headerbar should be saveable.
          */
         changeHandler = function() {
             setSaved.call(this, false);
         },
 
         /**
-         * sets dependent selects based on currently selected account
-         * @param id
-         * @param orderData
+         * @param {Object} contact
+         * @param {Object} orderData
+         */
+        initSelectsByContact = function(contact, orderData) {
+            var data = null;
+            var preselect = null;
+
+            // Load contact from repository.
+            var contactRepository = Contact.findOrCreate({id: contact.id});
+            contactRepository.fetch({
+                success: function(model) {
+                    var contactModel = model.toJSON();
+
+                    if (contactModel.hasOwnProperty('addresses')) {
+                        initAddressComponents.call(this, orderData, contactModel.addresses);
+                    }
+                }.bind(this),
+                error: function() {
+                    this.sandbox.emit('sulu.labels.warning.show', this.sandbox.translate('error while fetching account'));
+                }.bind(this)
+            });
+
+            // Load contact select.
+            if (!!contact && contact.id) {
+                data = [contact];
+                preselect = [contact.id];
+
+                initContactSelect.call(this, data, preselect);
+            }
+
+            initCustomerTypeSelect.call(this, CUSTOMER_TYPE.PRIVATE_PERSON);
+        },
+
+        /**
+         * Sets dependent selects based on currently selected account.
+         *
+         * @param {String} id
+         * @param {Object} orderData
          */
         initSelectsByAccountId = function(id, orderData) {
-            var data, preselect, account, addressData;
+            var data, preselect, account;
 
-            // load account
+            // Load account.
             account = Account.findOrCreate({id: id});
             account.fetch({
                 success: function(model) {
@@ -325,53 +619,16 @@ define([
                     }
 
                     if (account.hasOwnProperty('addresses')) {
-                        addressData = account.addresses;
-
-                        // when an address is already selected, the selected address should be used
-                        // otherwise the first delivery / payment address found will be used
-                        preselect = null;
-                        if (!orderData || !orderData.deliveryAddress) {
-                            preselect = findAddressWherePropertyIs.call(this, addressData, 'deliveryAddress', true);
-
-                            // when no delivery address is found the first address will be used
-                            if (!preselect && addressData.length > 0) {
-                                preselect = addressData[0];
-                            }
-                        } else {
-                            preselect = orderData.deliveryAddress;
-                        }
-                        this.sandbox.data.when(this.dfdDeliveryAddressInitialized).then(function() {
-                            initAddressComponents.call(this, addressData, constants.deliveryAddressInstanceName, preselect);
-                            setSettingsOverlayAdresses.call(this, addressData, preselect);
-                            this.options.data.deliveryAddress = preselect;
-                        }.bind(this));
-
-                        preselect = null;
-                        if (!orderData || !orderData.invoiceAddress) {
-                            preselect = findAddressWherePropertyIs.call(this, addressData, 'billingAddress', true);
-
-                            // when no invoice address is found the first address will be used
-                            if (!preselect && addressData.length > 0) {
-                                preselect = addressData[0];
-                            }
-                        } else {
-                            preselect = orderData.invoiceAddress;
-                        }
-
-                        this.sandbox.data.when(this.dfdInvoiceAddressInitialized).then(function() {
-                            initAddressComponents.call(this, addressData, constants.billingAddressInstanceName, preselect);
-                            this.options.data.invoiceAddress = preselect;
-                        }.bind(this));
+                        initAddressComponents.call(this, orderData, account.addresses);
                     }
 
                 }.bind(this),
                 error: function() {
-                    // TODO: MESSAGE
                     this.sandbox.emit('sulu.labels.warning.show', this.sandbox.translate('error while fetching account'));
                 }.bind(this)
             });
 
-            // load contacts of an account
+            // Load contacts of an account.
             this.sandbox.util.load(this.sandbox.util.template(constants.accountContactsUrl, {id: id}))
                 .then(function(response) {
                     data = response._embedded.contacts;
@@ -381,6 +638,8 @@ define([
                 .fail(function(textStatus, error) {
                     this.sandbox.logger.error(textStatus, error);
                 }.bind(this));
+
+            initCustomerTypeSelect.call(this, CUSTOMER_TYPE.ORGANIZATION);
         },
 
         /**
@@ -399,10 +658,11 @@ define([
         },
 
         /**
-         * Find an address where a specific property with a specific value is set
-         * @param addresses
-         * @param propertyName
-         * @param propertyValue
+         * Find an address where a specific property with a specific value is set.
+         *
+         * @param {Object} addresses
+         * @param {String} propertyName
+         * @param {String} propertyValue
          */
         findAddressWherePropertyIs = function(addresses, propertyName, propertyValue) {
             var address = null;
@@ -438,6 +698,7 @@ define([
             this.formId = form;
             this.accountId = null;
             this.contactId = null;
+            this.customerTypeId = null;
 
             this.dfdFormSaved = this.sandbox.data.deferred();
 
@@ -446,43 +707,47 @@ define([
             this.dfdShippingDate = this.sandbox.data.deferred();
             this.dfdOrderDate = this.sandbox.data.deferred();
 
+            this.dfdContactInitialized = this.sandbox.data.deferred();
+            this.dfdCustomerTypeInitialized = this.sandbox.data.deferred();
+
             this.dfdInvoiceAddressInitialized = this.sandbox.data.deferred();
             this.dfdDeliveryAddressInitialized = this.sandbox.data.deferred();
 
-            // define when all fields are initialized
+            // Define when all fields are initialized.
             this.sandbox.data.when(this.dfdShippingDate, this.dfdOrderDate, this.dfdAutoCompleteInitialized).then(function() {
                 this.dfdAllFieldsInitialized.resolve();
             }.bind(this));
 
             this.orderStatusId = getOrderStatusId.call(this);
 
-            // set if form is editable
+            // Set if form is editable.
             this.isEditable = this.orderStatusId <= OrderStatus.IN_CART;
 
             this.options.data = this.sandbox.util.extend({}, defaults, this.options.data);
 
-            // current id
+            // Current id.
             var id = this.options.data.id ? this.options.data.id : 'new';
 
-            this.accountInstanceName = 'customerAccount' + id;
+            this.customerAccountInstanceName = 'customerAccount' + id;
+            this.customerContactInstanceName = 'customerContact' + id;
 
-            // bind events
+            // Bind events.
             bindCustomEvents.call(this);
             bindDomEvents.call(this);
             HeaderUtil.initialize.call(this);
 
-            // set header
+            // Set header.
             setSaved.call(this, true);
 
-            // render form
+            // Render form.
             this.render();
 
             HeaderUtil.setToolbar.call(this, this.options.data);
 
-            // listen for changes in form
+            // Listen for changes in form.
             this.listenForChange();
 
-            // initialize sidebar
+            // Initialize sidebar.
             if (!!this.options.data && !!this.options.data.id && WidgetGroups.exists('order-detail')) {
                 Sidebar.initForDetail(this.sandbox, this.options.data);
             }
@@ -497,13 +762,13 @@ define([
 
             var data = this.options.data;
 
-            // initialize form
+            // Initialize form.
             initForm.call(this, data);
             this.startItemTableAndCurrencySelect();
         },
 
         /**
-         * Initializes the item-table and the select component
+         * Initializes the item-table and the select component.
          */
         startItemTableAndCurrencySelect: function() {
 
@@ -516,7 +781,7 @@ define([
                         remoteUrl: constants.accountUrl,
                         data: this.options.data.items,
                         currency: this.options.data.currencyCode,
-                        el: constants.itemTableSelector,
+                        el: selectors.itemTable,
                         enableIndependentItems: true,
                         settings: {
                             columns: [
@@ -543,7 +808,7 @@ define([
                 {
                     name: 'select@husky',
                     options: {
-                        el: constants.currencySelectSelector,
+                        el: selectors.currencySelect,
                         instanceName: constants.currencySelectInstanceName,
                         disabled: !this.isEditable,
                         emitValues: true,
@@ -569,34 +834,38 @@ define([
                     delete data.id;
                 }
 
-                // because the preselected option of the select conflicts with the data mapper
+                // Because the preselected option of the select conflicts with the data mapper
                 // the data mapper property is not used and therefore the data.currency property
-                // has to be set this way
+                // has to be set this way.
                 data.currencyCode = !!this.currency ? this.currency : this.options.data.currencyCode;
 
                 data.deliveryCost = this.deliveryCost;
 
-                // FIXME auto complete in mapper
-                // only get id, if auto-complete is not empty:
+                // Only get id, if auto-complete is not empty:
                 data.customerAccount = {
-                    id: this.sandbox.dom.attr('#' + this.accountInstanceName, 'data-id')
+                    id: getAccountId.call(this)
                 };
 
                 this.sandbox.logger.log('log data', data);
                 this.sandbox.emit('sulu.salesorder.order.save', data);
             } else {
-                this.sandbox.emit('sulu.labels.warning.show', this.sandbox.translate(constants.validateWarningTranslation));
+                this.sandbox.emit(
+                    'sulu.labels.warning.show',
+                    this.sandbox.translate(constants.validateWarningTranslation)
+                );
                 this.dfdFormSaved.reject();
             }
             return this.dfdFormSaved;
         },
 
-        // event listens for changes in form
+        /**
+         * Event listens for changes in form.
+         */
         listenForChange: function() {
-            // listen for change after TAGS and BIRTHDAY-field have been set
+            // Listen for change after TAGS and BIRTHDAY-field have been set.
             this.sandbox.data.when(this.dfdAllFieldsInitialized).then(function() {
 
-                // when input changes
+                // When input changes.
                 this.sandbox.dom.on(form, 'change', changeHandler.bind(this),
                     '.changeListener select, ' +
                     '.changeListener input, ' +
@@ -604,17 +873,14 @@ define([
                     '.changeListener .husky-select, ' +
                     '.changeListener textarea');
 
-                // on keyup
+                // On keyup.
                 this.sandbox.dom.on(form, 'keyup', changeHandler.bind(this),
                     '.changeListener select, ' +
                     '.changeListener input, ' +
                     '.changeListener textarea');
 
-                // change in item-table
+                // Change in item-table.
                 this.sandbox.on('sulu.item-table.changed', changeHandler.bind(this));
-
-                // TODO: use this for resetting account
-//                    this.sandbox.dom.on(constants.accountInputId+' input', 'changed', accountChangedListener.bind(this));
             }.bind(this));
 
         }